Update examples to Stack API
