+++ azure-cs-aks-private-container-registry/Azure.AKS.PrivateContainerRegistry.csproj
@@ -7,18 +7,18 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Pulumi.Azure" Version="1.4.0-preview"/>
-        <PackageReference Include="Pulumi.AzureAD" Version="1.5.0-preview"/>
-        <PackageReference Include="Pulumi.Docker" Version="1.2.0-preview"/>
-        <PackageReference Include="Pulumi.Kubernetes" Version="1.5.4-preview"/>
-        <PackageReference Include="Pulumi.Random" Version="1.5.0-preview"/>
-        <PackageReference Include="Pulumi.Tls" Version="1.4.0-preview"/>
+				<PackageReference Include="Pulumi.Azure" Version="2.1.0-preview" />
+				<PackageReference Include="Pulumi.AzureAD" Version="1.6.0-preview" />
+				<PackageReference Include="Pulumi.Docker" Version="1.4.0-preview" />
+				<PackageReference Include="Pulumi.Kubernetes" Version="1.5.8-preview" />
+				<PackageReference Include="Pulumi.Random" Version="1.6.0-preview" />
+				<PackageReference Include="Pulumi.Tls" Version="1.5.0-preview" />
     </ItemGroup>
 
     <ItemGroup>
-        <Compile Remove="SampleApplication\**"/>
-        <EmbeddedResource Remove="SampleApplication\**"/>
-        <None Remove="SampleApplication\**"/>
+        <Compile Remove="SampleApplication\**" />
+        <EmbeddedResource Remove="SampleApplication\**" />
+        <None Remove="SampleApplication\**" />
     </ItemGroup>
 
 </Project>
