+++ azure-cs-cosmosapp-component/Functions.cs
@@ -1,16 +1,11 @@
-﻿// Copyright 2016-2018, Pulumi Corporation.  All rights reserved.
-
-using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.IO;
+﻿// Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
 
 using Pulumi;
 using Pulumi.Azure.Core;
 
 public static class Functions
 {
-    public static IDictionary<string, object> Run()
+    public static Output<string> Run()
     {
         // Read a list of target locations from the config file:
         // Expecting a comma-separated list, e.g., "westus,eastus,westeurope"
@@ -44,9 +39,6 @@ public static class Functions
 
         });
 
-        return new Dictionary<string, object>
-        {
-            { "functionsEndpoint", Output.Format($"{app.Endpoint}/cosmos") }
-        };
+        return Output.Format($"{app.Endpoint}/cosmos");
     }
 }
