+++ azure-cs-msi-keyvault-rbac/README.md
@@ -66,9 +66,9 @@ To deploy your infrastructure, follow the below steps.
 1.  Check the deployed website endpoint:
 
     ```
-    $ pulumi stack output endpoint
+    $ pulumi stack output Endpoint
     https://app129968b8.azurewebsites.net/
-    $ curl "$(pulumi stack output endpoint)"
+    $ curl "$(pulumi stack output Endpoint)"
     Hello 311378b3-16b7-4889-a8d7-2eb77478beba@50f73f6a-e8e3-46b6-969c-bf026712a650! Here is your...
     ```
