+++ digitalocean-cs-k8s/Digitalocean.K8s.csproj
@@ -9,8 +9,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Pulumi.Digitalocean" Version="1.2.0-preview" />
-    <PackageReference Include="Pulumi.Kubernetes" Version="1.3.3-preview" />
+    <PackageReference Include="Pulumi.DigitalOcean" Version="1.7.0-preview" />
+    <PackageReference Include="Pulumi.Kubernetes" Version="1.5.8-preview" />
   </ItemGroup>
 
 </Project>
