+++ gcp-cs-gke/README.md
@@ -82,8 +82,8 @@ After cloning this repo, from this working directory, run these commands:
     +   └─ gcp:container:Cluster  helloworld      created
 
     Outputs:
-        clusterName: "helloworld-9b9530f"
-        kubeconfig : "<KUBECONFIG_CONTENTS>"
+        ClusterName: "helloworld-9b9530f"
+        KubeConfig : "<KUBECONFIG_CONTENTS>"
 
 	Resources:
         + 2 created
@@ -101,7 +101,7 @@ After cloning this repo, from this working directory, run these commands:
     stack output in the CLI, as Pulumi faciliates exporting these objects for us.
 
     ```bash
-    $ pulumi stack output kubeconfig > kubeconfig
+    $ pulumi stack output KubeConfig > kubeconfig
     $ export KUBECONFIG=$PWD/kubeconfig
 
     $ kubectl version
